tools/requirements.txt
streamlit>=1.30.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
openpyxl>=3.1.0
openai>=1.6.0
